+++ agent/agent_configuration.go
@@ -30,5 +30,5 @@ type AgentConfiguration struct {
 	Profile                    string
 	RedactedVars               []string
 	AcquireJob                 string
-	TracingDatadogAddr         string
+	TracingBackend             string
 }